backend/requirements.txt
fastapi
uvicorn
pydantic
faiss-cpu
sentence-transformers
huggingface-hub
numpy
